factor: improve table code factoring
